+++ packages/python/qsu/web/__init__.py
@@ -1,5 +1,6 @@
 from .generateLicense import generateLicense
 from .getParsedInfoFromAddress import getParsedInfoFromAddress
+from .getSlug import getSlug
 from .isBotAgent import isBotAgent
 from .isMatchPathname import isMatchPathname
 from .isMobile import isMobile
@@ -8,6 +9,7 @@ from .removeLocalePrefix import removeLocalePrefix
 __all__ = [
 	'generateLicense',
 	'getParsedInfoFromAddress',
+	'getSlug',
 	'isBotAgent',
 	'isMatchPathname',
 	'isMobile',